퀴즈 문제 답안
현재시간+삭제시 패스워드 처리
(패스워는 1234)

+++ ex02/src/main/java/org/zerock/controller/BoardController.java
@@ -27,6 +27,7 @@ public class BoardController {
 		log.info("list url 요청..");
 		model.addAttribute("list", service.getList(cri)); //글목록
 		model.addAttribute("pageMaker", new PageDTO(cri, service.count(cri)));    //페이지바 정보
+		model.addAttribute("now", service.now());
 		// -> board/list.jsp
 	}
 	
@@ -41,9 +42,14 @@ public class BoardController {
 									   //redirect 가 있으면 요청
 	}
 	
-	@GetMapping({"/register","/remove"})
+	@GetMapping("/register")
 	public void register() {
 		
+	}
+	
+	@GetMapping("/remove") //비번 화면 요청
+	public void remove(Long bno,Model model) {
+		model.addAttribute("bno", bno);
 	}
 
 	@GetMapping("/modify")
@@ -66,13 +72,19 @@ public class BoardController {
 	
 	// 삭제(글번호-bno) board/remove (post)  <- 입력화면(get)
 	@PostMapping("/remove")
-	public String remove(Long bno,RedirectAttributes rttr) {
+	public String remove(Long bno,RedirectAttributes rttr,String pw) {
 		log.info("삭제 url 요청");
-		if(service.remove(bno)) { //이상없으면 result 이름으로 success 라는 문자 전송
-			rttr.addFlashAttribute("oper", "remove");
-			rttr.addFlashAttribute("result", bno); 
+		log.info("입력된 패스워드:"+ pw);
+		if(pw.equals("1234")) { 
+			if(service.remove(bno)) { //이상없으면 result 이름으로 success 라는 문자 전송
+				rttr.addFlashAttribute("oper", "remove");
+				rttr.addFlashAttribute("result", bno); 
+			}
+			return "redirect:/board/list";	
+		}else {
+			rttr.addFlashAttribute("flag", "fail");
+			return "redirect:/board/remove?bno="+bno;
 		}
-		return "redirect:/board/list";	
 	}
 
 	
@@ -80,12 +92,14 @@ public class BoardController {
 	@PostMapping("/modify")
 	public String modify(BoardVO vo,RedirectAttributes rttr,Criteria cri) {
 		log.info("수정 url 요청:");
+	
 		if(service.modify(vo)) {
 			rttr.addFlashAttribute("oper", "modify");
 			rttr.addFlashAttribute("result", vo.getBno()); 
 		}
 		return "redirect:/board/list?pageNum="+cri.getPageNum()+
 				"&amount="+cri.getAmount();	
+	
 	}
 	
 	//좋아요 처리
